+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0 (February 2020)
+* Added support for Remote Commands
+* Underlying Tealium libraries updated to versions 5.7.0 (Android) and 5.6.6 (iOS) 
+
 ## 1.0.1 (August 2019)
 * Update documentation link